+++ Code/factory/__init__.py
@@ -1,3 +1,4 @@
 from factory.base_component import Component
 from factory.base_entity import Entity
-from factory.prototype_factory import PrototypeError, PrototypeFactory
+from factory.errors import PrototypeError
+from factory.prototype_factory import PrototypeFactory
