--- Docs/base_classes/base_component.md
@@ -1,19 +1,0 @@
-# Документация по файлу `base_component.py`
-
-
-## `Component.__init__`<br>
-Инициализирует базовый объект компонента.<br>
-<br>
-
-## `Component.owner`<br>
-Возвращает родительскую сущьность для компонента<br>
-**Returns:**<br>
-Entity: Родительская сущность.<br>
-Устанавливает родительскую сущность для компонента.<br>
-**Args:**<br>
-entity (Entity): Родительская сущность.<br>
-<br>
-
-## `Component.update`<br>
-Обновляет состояние компонента.<br>
-<br>
