--- Docs/base_classes/base_object.md
@@ -1,25 +1,0 @@
-# Документация по файлу `base_object.py`
-
-
-## `BaseObject.__init__`<br>
-Инициализация класса базового объекта. Включает в себя базовую логику работы с базовыми переменными<br>
-**Attributes:**<br>
-id (str): ID объекта<br>
-name (str): Имя объекта<br>
-description (str): Описание объекта<br>
-<br>
-
-## `BaseObject.id`<br>
-Публичное поле id объекта<br>
-Является одновременно сеттером и геттером<br>
-<br>
-
-## `BaseObject.name`<br>
-Публичное поле name объекта<br>
-Является одновременно сеттером и геттером<br>
-<br>
-
-## `BaseObject.description`<br>
-Публичное поле description объекта<br>
-Является одновременно сеттером и геттером<br>
-<br>
