--- Docs/medical/need.md
@@ -1,64 +1,0 @@
-# Документация по файлу `need.py`
-
-
-## `Need.__init__`<br>
-Инициализация объекта Need.<br>
-Наследуется от BaseObject.<br>
-**Attributes:**<br>
-value (int): Текущее значение потребности. <br>
-max_value (int): Максимальное значение потребности. <br>
-min_value (int): Минимальное значение потребности.<br>
-count (int, optional): Дельта изменения за ход/тик. По умолчанию -1.<br>
-<br>
-
-## `Need.value`<br>
-Получение текущего значения потребности.<br>
-**Returns:**<br>
-int: Текущее значение потребности.<br>
-<br>
-Установка текущего значения потребности.<br>
-**Args:**<br>
-value (int): Текущее значение потребности.<br>
-<br>
-
-## `Need.max_value`<br>
-Получение максимального значения потребности.<br>
-**Returns:**<br>
-int: Максимальное значение потребности.<br>
-<br>
-Установка максимального значения потребности.<br>
-**Args:**<br>
-max_value (int): Максимальное значение потребности.<br>
-<br>
-
-## `Need.min_value`<br>
-Получение минимального значения потребности.<br>
-**Returns:**<br>
-int: Минимальное значение потребности.<br>
-<br>
-Установка минимального значения потребности.<br>
-**Args:**<br>
-min_value (int): Минимальное значение потребности.<br>
-<br>
-
-## `Need.count`<br>
-Получение дельты изменения за ход/тик.<br>
-**Returns:**<br>
-int: Дельта изменения за ход/тик.<br>
-<br>
-
-## `Need.set_count`<br>
-Установка дельты изменения за ход/тик.<br>
-**Args:**<br>
-count (int): Дельта изменения за ход/тик.<br>
-<br>
-
-## `Need.update`<br>
-Изменение текущего значения потребности на значение дельты изменения за ход/тик.<br>
-<br>
-
-## `Need.value_change`<br>
-Изменение текущего значения потребности на указанное число.<br>
-**Args:**<br>
-value (int): Число, на которое изменяется текущее значение потребности.<br>
-<br>
